Move the Newsletter class into the service package
