+++ src/test/scala/com/gu/newsletterlistcleanse/services/NewslettersSpec.scala
@@ -1,4 +1,4 @@
-package com.gu.newsletterlistcleanse
+package com.gu.newsletterlistcleanse.services
 
 import java.time.ZonedDateTime
